requirements.txt
certifi==2024.7.4
charset-normalizer==3.1.0
decorator==5.1.1
idna==3.7
oauthlib==3.2.2
py==1.11.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.32.4
requests-oauthlib==1.3.1
retry==0.9.2
six==1.16.0
TeslaPy==2.8.0
urllib3==2.6.0
websocket-client==1.5.2

# New dependencies for Home Assistant integration
paho-mqtt==2.1.0
python-dotenv==1.0.0
PyYAML==6.0.1

# InfluxDB for historical data storage
influxdb-client==1.38.0

# Sun position calculations for solar optimization
astral==3.2
